% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_import.R
\name{from_import}
\alias{from_import}
\title{from_import()}
\usage{
from_import(package, ..., verbose = TRUE)
}
\arguments{
\item{package}{the package name (length one character vector)}

\item{...}{one or more function names (as-is, or as a list)}

\item{verbose}{give a warning when overwriting existing objects? TRUE by
default.}
}
\description{
Imports one or more functions from a package into the parent environment,
overwriting any existing objects in the parent environment sharing the name
of the function(s). The package is loaded but not attached to the search
list. Inspired by Python.
}
\examples{
\dontrun{
from_import("dplyr", "select", "filter")
from_import("dplyr", list("select", "filter"))
}
}

