% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{install_package}
\alias{install_package}
\title{install_package()
Function to install a package given a package object}
\usage{
install_package(pack)
}
\arguments{
\item{pack}{a \code{package_obj} object}
}
\description{
install_package()
Function to install a package given a package object
}

