% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_packages.R
\name{install_packages}
\alias{install_packages}
\title{install_packages()}
\usage{
install_packages(...)
}
\arguments{
\item{...}{one or more package names or \code{package_obj} objects}
}
\description{
Installs one or more packages. Similar to \code{utils::install.packages}, but
you may supply either package names and \code{package_obj} objects to this
function. You can install Github packages by supplying \code{username/repo}
to this function, or \code{username$repo} for Bitbucket packages.
}
\details{
\code{package_obj} allows you to supply it an install function if
the package isn't on CRAN or in a public GitHub or Bitbucket repo.
}
\examples{
\dontrun{
install_packages("dplyr", "ggplot2", "rvest", "magrittr")
}
}

