% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{is.package_obj}
\alias{is.package_obj}
\title{is.package_obj()
Returns TRUE if obj is a package_obj object}
\usage{
is.package_obj(obj)
}
\arguments{
\item{obj}{an R object}
}
\description{
is.package_obj()
Returns TRUE if obj is a package_obj object
}

