% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraries.R
\name{libraries}
\alias{libraries}
\title{libraries()}
\usage{
libraries(...)
}
\arguments{
\item{...}{one or more package names or \code{package_obj} objects}
}
\description{
A vectorized version of the \code{library} function that accepts one or
more package names to call \code{library} on. Unlike \code{library},
the \code{libraries} function does not use non-standard evaluation (NSE)
on its inputs, meaning that you must supply your package names as strings,
or as variables.
}
\details{
You may supply either package names and \code{package_obj} objects to this
function. You may also add \code{::} to the end of a package name to load
the package instead of attaching the package (this means that the package
will not be added to the search list, but will be available to access via
the \code{::} operator).
}
\examples{
\dontrun{
libraries("dplyr", "ggplot2", "rvest", "magrittr")
libraries("dplyr::", "Rdatatable/data.table")
}
}

