% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{load_package}
\alias{load_package}
\title{load_package()
Function to load a package given a package object}
\usage{
load_package(pack)
}
\arguments{
\item{pack}{a \code{package_obj} object}
}
\description{
load_package()
Function to load a package given a package object
}

